Creating screens folder and moving homescreen in new homescreen class
